refactor(d3): split 6 renderers into per-viz modules (109.2 step 4)
Each visualization renderer + its private helpers now lives in its own module:
src/{slope,arc,timeline,wordcloud,venn,quadrant}/renderer.ts. The blocks were
fully self-contained (no cross-viz refs) once the shared tooltip helpers
(showTooltip/hideTooltip) joined d3-helpers. d3.ts collapses to a 1157-line
export-dispatch + public re-export barrel (was 8863). Renderers stay fat-typed
on ParsedVisualization for now; narrow discriminated union lands next.

Zero snapshot drift, 6831 tests pass, typecheck clean, 5 cycles.
